+++ gateway/src/apicast/configuration/service.lua
@@ -7,7 +7,12 @@ local tostring = tostring
 local rawget = rawget
 local lower = string.lower
 local gsub = string.gsub
+local format = string.format
 local select = select
+local concat = table.concat
+local insert = table.insert
+local re = require 'ngx.re'
+local re_match = ngx.re.match
 
 local http_authorization = require 'resty.http_authorization'
 
@@ -160,6 +165,58 @@ function backend_version_credentials.version_oauth(config)
   return setmetatable({ access_token, access_token = access_token }, credentials_oauth_mt)
 end
 
+local function set_or_inc(t, name, delta)
+  return (t[name] or 0) + (delta or 0)
+end
+
+local function check_rule(req, rule, usage_t, matched_rules, params)
+  local pattern = rule.regexpified_pattern
+  local match = re_match(req.path, format("^%s", pattern), 'oj')
+
+  if match and req.method == rule.method then
+    local args = req.args
+
+    if rule.querystring_params(args) then -- may return an empty table
+      local system_name = rule.system_name
+      -- FIXME: this had no effect, what is it supposed to do?
+      -- when no querystringparams
+      -- in the rule. it's fine
+      -- for i,p in ipairs(rule.parameters or {}) do
+      --   param[p] = match[i]
+      -- end
+
+      local value = set_or_inc(usage_t, system_name, rule.delta)
+
+      usage_t[system_name] = value
+      params['usage[' .. system_name .. ']'] = value
+      insert(matched_rules, rule.pattern)
+    end
+  end
+end
+
+local function get_auth_params(method)
+  local params = ngx.req.get_uri_args()
+
+  if method == "GET" then
+    return params
+  else
+    ngx.req.read_body()
+    local body_params, err = ngx.req.get_post_args()
+
+    if not body_params then
+      ngx.log(ngx.NOTICE, 'Error while getting post args: ', err)
+      body_params = {}
+    end
+
+    -- Adds to body_params URI params that are not included in the body. Doing
+    -- the reverse would be more expensive, because in general, we expect the
+    -- size of body_params to be larger than the size of params.
+    setmetatable(body_params, { __index = params })
+
+    return body_params
+  end
+end
+
 -- This table can be used with `table.concat` to serialize
 -- just the numeric keys, but also with `pairs` to iterate
 -- over just the non numeric keys (for query building).
@@ -193,6 +250,52 @@ function _M:oauth()
     return oauth.apicast.new(self)
   else
     return nil, 'not oauth'
+  end
+end
+
+local function extract_usage_v2(config, method, path)
+  local usage_t =  {}
+  local matched_rules = {}
+  local params = {}
+  local rules = config.rules
+
+  local args = get_auth_params(method)
+
+  ngx.log(ngx.DEBUG, '[mapping] service ', config.id, ' has ', #rules, ' rules')
+
+  for i = 1, #rules do
+    check_rule({path=path, method=method, args=args}, rules[i], usage_t, matched_rules, params)
+  end
+
+  -- if there was no match, usage is set to nil and it will respond a 404, this
+  -- behavior can be changed
+  return usage_t, concat(matched_rules, ", "), params
+end
+
+-- Deprecated
+function _M:extract_usage(request)
+  ngx.log(ngx.WARN, 'extract_usage is deprecated, please use get_usage(method, path)')
+  local req = re.split(request, " ", 'oj')
+  local method, url = req[1], req[2]
+  local path = re.split(url, "\\?", 'oj')[1]
+
+  return extract_usage_v2(self, method, path)
+end
+
+--- Get the usage associated with a request
+-- @tparam string method Method of the request (GET, POST, etc.)
+-- @tparam string path Path of the request
+function _M:get_usage(method, path)
+  -- This is a simple dispatcher. If it detects that the 'extract_usage' method
+  -- has been defined, it calls it. Otherwise, it calls the new version of the
+  -- method. This is done to keep backwards compatibility, because in previous
+  -- versions it was possible to ovewrite that method and expect to be called
+  -- from where get_usage is currently being called.
+
+  if self.extract_usage and self.extract_usage ~= _M.extract_usage then
+    return self:extract_usage(ngx.var.request)
+  else
+    return extract_usage_v2(self, method, path)
   end
 end